requirements.txt
numpy==1.26.4
Flask==3.1.3
GitPython==3.1.46
networkx==3.6.1
pymongo==4.16.0
python-dotenv==1.2.2
requests==2.32.5
scikit-learn==1.8.0
torch==2.11.0
transformers==4.41.2
datasets==2.19.1
accelerate==0.30.1